Move common into pkg/common
