--- common/Makefile
@@ -1,10 +1,0 @@
-include ../common.mk
-include ../help.mk
-
-##@ Development
-
-fmt: ## Run go fmt against code.
-	go fmt ./...
-
-vet: ## Run go vet against code.
-	go vet ./...
