+++ controllers/pipelines/internal/workflowfactory/experiment_workflow_integration_test.go
@@ -6,8 +6,8 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	config "github.com/sky-uk/kfp-operator/apis/config/hub"
 	pipelineshub "github.com/sky-uk/kfp-operator/apis/pipelines/hub"
-	"github.com/sky-uk/kfp-operator/argo/providers/base"
-	testutil "github.com/sky-uk/kfp-operator/common/testutil/provider"
+	testutil "github.com/sky-uk/kfp-operator/pkg/common/testutil/provider"
+	"github.com/sky-uk/kfp-operator/pkg/providers/base"
 )
 
 var _ = Context("Experiment Resource Workflows", Serial, func() {
