+++ controllers/pipelines/internal/workflowfactory/run_workflow_integration_test.go
@@ -8,8 +8,8 @@ import (
 	. "github.com/onsi/gomega"
 	config "github.com/sky-uk/kfp-operator/apis/config/hub"
 	pipelineshub "github.com/sky-uk/kfp-operator/apis/pipelines/hub"
-	testutil "github.com/sky-uk/kfp-operator/common/testutil/provider"
 	. "github.com/sky-uk/kfp-operator/controllers/pipelines/internal/testutil"
+	testutil "github.com/sky-uk/kfp-operator/pkg/common/testutil/provider"
 	"github.com/sky-uk/kfp-operator/pkg/providers/base"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
