+++ main.go
@@ -23,8 +23,8 @@ import (
 	"os"
 
 	"github.com/samber/lo"
-	"github.com/sky-uk/kfp-operator/common/metrics"
 	"github.com/sky-uk/kfp-operator/controllers/webhook"
+	"github.com/sky-uk/kfp-operator/pkg/common/metrics"
 	"k8s.io/apimachinery/pkg/util/yaml"
 	runtimeMetrics "sigs.k8s.io/controller-runtime/pkg/metrics"
