+++ pkg/common/triggers/indicator.go
