+++ pkg/common/triggers/indicator_test.go
