+++ provider-service/vai/internal/provider/label_gen.go
@@ -3,8 +3,8 @@ package provider
 import (
 	"fmt"
 
-	"github.com/sky-uk/kfp-operator/common/triggers"
 	"github.com/sky-uk/kfp-operator/pkg/common"
+	"github.com/sky-uk/kfp-operator/pkg/common/triggers"
 	"github.com/sky-uk/kfp-operator/pkg/providers/base"
 	"github.com/sky-uk/kfp-operator/provider-service/vai/internal/label"
 )
